Put RangeBox into its own component
